examples/card_manager_app/requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
werkzeug==3.0.1
Jinja2==3.1.2
